LayoutTests/platform/ios-sim/svg/text/lengthAdjust-text-metrics-expected.txt

Complete test of the SVGTextContentElement API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test SVGTextContentElement SVG DOM properties
PASS svgText.textLength.baseVal.value.toFixed(1) is "200.0"
PASS svgText.lengthAdjust.baseVal is SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS

Test getNumberOfChars() API
PASS svgText.getNumberOfChars() is 4

Test getComputedTextLength() API
PASS lengthToString(svgText.getComputedTextLength()) is "200.0"

Test getSubStringLength() API
FAIL lengthToString(svgText.getSubStringLength(0, 1)) should be 61.5. Was 62.8.
FAIL lengthToString(svgText.getSubStringLength(0, 2)) should be 117.9. Was 120.0.
FAIL lengthToString(svgText.getSubStringLength(0, 3)) should be 169.2. Was 171.4.
PASS lengthToString(svgText.getSubStringLength(0, 4)) is "200.0"
FAIL lengthToString(svgText.getSubStringLength(1, 1)) should be 56.4. Was 57.2.
FAIL lengthToString(svgText.getSubStringLength(1, 2)) should be 107.7. Was 108.6.
FAIL lengthToString(svgText.getSubStringLength(1, 3)) should be 138.5. Was 137.2.
FAIL lengthToString(svgText.getSubStringLength(2, 1)) should be 51.3. Was 51.4.
FAIL lengthToString(svgText.getSubStringLength(2, 2)) should be 82.1. Was 80.0.
FAIL lengthToString(svgText.getSubStringLength(3, 1)) should be 30.8. Was 28.6.

Test getStartPositionOfChar() API
PASS pointToString(svgText.getStartPositionOfChar(0)) is "(10.0,20.0)"
FAIL pointToString(svgText.getStartPositionOfChar(1)) should be (71.5,20.0). Was (72.8,20.0).
FAIL pointToString(svgText.getStartPositionOfChar(2)) should be (127.9,20.0). Was (130.0,20.0).
FAIL pointToString(svgText.getStartPositionOfChar(3)) should be (179.2,20.0). Was (181.4,20.0).

Test getEndPositionOfChar() API
FAIL pointToString(svgText.getEndPositionOfChar(0)) should be (71.5,20.0). Was (72.8,20.0).
FAIL pointToString(svgText.getEndPositionOfChar(1)) should be (127.9,20.0). Was (130.0,20.0).
FAIL pointToString(svgText.getEndPositionOfChar(2)) should be (179.2,20.0). Was (181.4,20.0).
PASS pointToString(svgText.getEndPositionOfChar(3)) is "(210.0,20.0)"

Test getExtentOfChar() API
FAIL rectToString(svgText.getExtentOfChar(0)) should be (10.0,1.9)-(61.5x22.3). Was (10.0,1.0)-(62.8x24.0).
FAIL rectToString(svgText.getExtentOfChar(1)) should be (71.5,1.9)-(56.4x22.3). Was (72.8,1.0)-(57.2x24.0).
FAIL rectToString(svgText.getExtentOfChar(2)) should be (127.9,1.9)-(51.3x22.3). Was (130.0,1.0)-(51.4x24.0).
FAIL rectToString(svgText.getExtentOfChar(3)) should be (179.2,1.9)-(30.8x22.3). Was (181.4,1.0)-(28.6x24.0).

Test getRotationOfChar() API
PASS svgText.getRotationOfChar(0).toFixed(1) is "0.0"
PASS svgText.getRotationOfChar(1).toFixed(1) is "0.0"
PASS svgText.getRotationOfChar(2).toFixed(1) is "0.0"
PASS svgText.getRotationOfChar(3).toFixed(1) is "0.0"

Test getCharNumAtPosition() API
> Testing point=(0.0,10.0)
PASS svgText.getCharNumAtPosition(point) is -1
> Testing point=(9.9,10.0)
PASS svgText.getCharNumAtPosition(point) is -1
> Testing point=(10.1,10.0)
PASS svgText.getCharNumAtPosition(point) is 0
> Testing point=(71.4,10.0)
PASS svgText.getCharNumAtPosition(point) is 0
> Testing point=(71.6,10.0)
FAIL svgText.getCharNumAtPosition(point) should be 1. Was 0.
> Testing point=(127.8,10.0)
PASS svgText.getCharNumAtPosition(point) is 1
> Testing point=(128.0,10.0)
FAIL svgText.getCharNumAtPosition(point) should be 2. Was 1.
> Testing point=(179.1,10.0)
PASS svgText.getCharNumAtPosition(point) is 2
> Testing point=(179.3,10.0)
FAIL svgText.getCharNumAtPosition(point) should be 3. Was 2.
> Testing point=(209.9,10.0)
PASS svgText.getCharNumAtPosition(point) is 3
> Testing point=(210.1,10.0)
PASS svgText.getCharNumAtPosition(point) is -1
> Testing point=(250.0,10.0)
PASS svgText.getCharNumAtPosition(point) is -1
PASS successfullyParsed is true

TEST COMPLETE
